pyproject.toml
[tool.poetry]
name = "to-service"
version = "0.1.0"
description = "Data engineering project to visualize frequency of Toronto service requests"
authors = ["vykuang <vykuang92@gmail.com>"]
readme = "README.md"
packages = [{include = "to_service"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
prefect = "^2.8"
pyarrow = "^10.0.1"
gcsfs = "^2023.1.0"
google-cloud-bigquery = "^3.8.0"
prefect-dbt = "^0.3.1"
dbt-bigquery = "^1.5.1"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.1.0"}
pytest = "^7.2.2"
jupyterlab = "^3.6.1"
ipykernel = "^6.21.3"
pre-commit = "^3.2.0"
python-dotenv = "^1.0.0"
folium = "^0.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
